[breaking] use instafeed-lite, create Wrapper Component
Not a good commit! Haha, I apologize for this, however, this is more of
a
 foundation to get this started again.

This is now only a React Wrapper Component that will call
`instafeed-lite`,
 which only returns an Object now.

We simply need to pass the parameters we get, to `instafeed-lite` then
wrap
 it in a React Component, and go away.

Again ... there is really not a true point to this Repo in React.
This can easily be done by doing a `fetch` on your own in either:
- componentDidMount
- getInitialProps
Updated README to reflect that.